Update activity type provider to enable creating

+++ src/context/activityType/activityTypeProvider.tsx
@@ -7,12 +7,12 @@ import AgendaContext from '../AgendaContext';
 import { CurrentEventContext } from '../eventContext';
 
 import ActivityTypeContext from './activityTypeContext';
-import { ActivityTypeValueType } from './schema/structureInterfaces';
+import { ActivitySubTypeValueType, ActivityTypeValueType } from './schema/structureInterfaces';
 import {
   ActivityTypeProviderProps,
   ActivityTypeContextType,
 } from './types/types';
-import { activityTypeData } from './schema/activityTypeFormStructure';
+import { activitySubTypeKeys, activityTypeData } from './schema/activityTypeFormStructure';
 // Temporally
 import { ExtendedAgendaDocumentType } from '@/components/agenda/types/AgendaDocumentType';
 
@@ -26,7 +26,7 @@ function ActivityTypeProvider(props: ActivityTypeProviderProps) {
     activityName,
     activityDispatch,
     dataLive,
-    meeting_id,
+    meeting_id: meetingId,
     activityEdit,
     setDataLive,
     setHabilitarIngreso,
@@ -37,17 +37,29 @@ function ActivityTypeProvider(props: ActivityTypeProviderProps) {
   const [isSavingActivityType, setIsSavingActivityType] = useState(false);
   const [isDeletingActivityType, setIsDeletingActivityType] = useState(false);
   const [isUpdatingAcctivityType, setIsUpdatingAcctivityType] = useState(false);
+  const [isUpdatingActivityContent, setIsUpdatingActivityContent] = useState(false);
   const [videoObject, setVideoObject] = useState<any | null>(null);
   const [activityType, setActivityType] = useState<ActivityTypeValueType | null>(null);
+  const [activityContentType, setActivityContentType] = useState<ActivitySubTypeValueType | null>(null);
+  const [contentSource, setContentSource] = useState<string | null>(null);
 
   const queryClient = useQueryClient();
 
   const saveActivityType = () => {
     console.debug('activity type provider is saving...');
     console.debug('activityType is:', activityType);
-    if (activityType === null) return;
+    if (activityType === null) {
+      console.error('activityType (from ActivityTypeProvider) is none');
+      return;
+    }
+
     if (!(cEvent?.value?._id)) {
       console.error('ActivityTypeProvider.saveActivityType cannot get cEvent.value._id');
+      return;
+    }
+
+    if (!activityEdit) {
+      console.error('activityEdit (from AgendaContext) is none');
       return;
     }
 
@@ -75,12 +87,20 @@ function ActivityTypeProvider(props: ActivityTypeProviderProps) {
 
   const deleteActivityType = async () => {
     if (!(cEvent?.value?._id)) {
-      console.error('ActivityTypeProvider.saveActivityType cannot get cEvent.value._id');
+      console.error('ActivityTypeProvider.deleteActivityType cannot get cEvent.value._id');
       return;
     }
+
+    if (!activityEdit) {
+      console.error('activityEdit (from AgendaContext) is none');
+      return;
+    }
+
     console.debug('AT provider is deleting');
 
     setIsDeletingActivityType(true);
+
+    setContentSource(null);
     try {
       await TypesAgendaApi.deleteOne(activityEdit, cEvent.value._id);
       console.debug('AT provider delete successfully');
@@ -92,6 +112,127 @@ function ActivityTypeProvider(props: ActivityTypeProviderProps) {
     }
   };
 
+  const saveActivityContent = () => {
+    if (activityType === null) {
+      console.error('activityType (from ActivityTypeProvider) is none');
+      return;
+    }
+
+    if (!(cEvent?.value?._id)) {
+      console.error('ActivityTypeProvider.saveActivityContent cannot get cEvent.value._id');
+      return;
+    }
+
+    if (!activityEdit) {
+      console.error('activityEdit (from AgendaContext) is none');
+      return;
+    }
+
+    setIsUpdatingActivityContent(true);
+    (async () => {
+      switch (activityContentType) {
+        case activitySubTypeKeys.url: {
+          const respUrl = await AgendaApi.editOne({ video: contentSource }, activityEdit, cEvent.value._id);
+          if (respUrl) {
+            await saveConfig({
+              platformNew: '',
+              type: activitySubTypeKeys.url,
+              habilitar_ingreso: '',
+              data: contentSource,
+            });
+            setTypeActivity(activitySubTypeKeys.url);
+            setPlatform('wowza');
+            setMeetingId(contentSource);
+          }
+          break;
+        }
+        case activitySubTypeKeys.vimeo: {
+          const resp = await saveConfig({ platformNew: 'vimeo', type: 'vimeo', data: contentSource });
+          setTypeActivity(activitySubTypeKeys.vimeo);
+          setPlatform(activitySubTypeKeys.vimeo);
+          setMeetingId(contentSource);
+          break;
+        }
+        case activitySubTypeKeys.youtube: {
+          if (!contentSource) {
+            console.error('ActivityTypeProvider: contentSource is none');
+            return;
+          }
+          let newData = contentSource.includes('https://youtu.be/')
+            ? contentSource
+            : 'https://youtu.be/' + contentSource;
+          const resp = await saveConfig({ platformNew: 'wowza', type: activitySubTypeKeys.youtube, data: newData });
+          setTypeActivity('youTube');
+          setPlatform('wowza');
+          setMeetingId(contentSource);
+          break;
+        }
+        case activitySubTypeKeys.meeting: {
+          const resp = await saveConfig({
+            platformNew: '',
+            type: activitySubTypeKeys.meeting,
+            data: contentSource,
+            habilitar_ingreso: 'only',
+          });
+          setTypeActivity(activitySubTypeKeys.meeting);
+          setPlatform('wowza');
+          break;
+        }
+        case activitySubTypeKeys.file: {
+          if (!contentSource) {
+            console.error('ActivityTypeProvider: contentSource is none');
+            return;
+          }
+          const data = contentSource.split('*');
+          const urlVideo = data[0];
+          const respUrlVideo = await AgendaApi.editOne({ video: urlVideo }, activityEdit, cEvent.value._id);
+          if (respUrlVideo) {
+            const resp = await saveConfig({ platformNew: '', type: 'video', data: urlVideo, habilitar_ingreso: '' });
+            setTypeActivity('video');
+            setPlatform('wowza');
+            setMeetingId(urlVideo);
+          }
+          break;
+        }
+        case 'eviusMeet': {
+          !meetingId && executer_createStream.mutate();
+          meetingId &&
+            (await saveConfig({
+              platformNew: 'wowza',
+              type: activityContentType,
+              data: meetingId,
+            }));
+          setTypeActivity('eviusMeet');
+          setPlatform('wowza');
+          break;
+        }
+        case 'RTMP': {
+          !meetingId && executer_createStream.mutate();
+          meetingId &&
+            (await saveConfig({ platformNew: 'wowza', type: activityContentType, data: meetingId }));
+          setTypeActivity('RTMP');
+          setPlatform('wowza');
+          break;
+        }
+        default:
+          // alert(`wtf is ${activityContentType}`);
+          console.warn(`wtf is ${activityContentType}`);
+      }
+    })();
+  };
+
+  const executer_createStream = useMutation(() => createLiveStream(activityName), {
+    onSuccess: async (data: any) => {
+      queryClient.setQueryData('livestream', data);
+
+      await saveConfig({ platformNew: 'wowza', type: activityContentType, data: data.id });
+      setDataLive(data);
+      activityDispatch({ type: 'meeting_created', meeting_id: data.id });
+      // Invalidate and refetch
+      //queryClient.invalidateQueries('todos')
+    },
+  });
+
   const value: ActivityTypeContextType = {
     // Flags
     is: {
@@ -100,15 +241,21 @@ function ActivityTypeProvider(props: ActivityTypeProviderProps) {
       deleting: isDeletingActivityType,
       saving: isSavingActivityType,
       updatingActivityType: isUpdatingAcctivityType,
+      updatingActivityContent: isUpdatingActivityContent,
     },
     // Objects
     formWidgetFlow: activityTypeData,
     videoObject,
     activityType,
+    contentSource,
+    activityContentType,
     // Functions
     setActivityType,
     saveActivityType,
     deleteActivityType,
+    setContentSource,
+    saveActivityContent,
+    setActivityContentType,
   };
 
   useEffect(() => {
@@ -123,8 +270,25 @@ function ActivityTypeProvider(props: ActivityTypeProviderProps) {
         const agendaInfo: ExtendedAgendaDocumentType = await AgendaApi
           .getOne(activityEdit, cEvent.value._id);
         // setDefinedType(agendaInfo.type?.name || null);
-        const typeIncomming = agendaInfo.type?.name as ActivityTypeValueType;
-        if (typeIncomming) setActivityType(typeIncomming);
+        const typeIncoming = agendaInfo.type?.name as ActivityTypeValueType;
+
+        const onlyActivityTypes: ActivityTypeValueType[] = [
+          'liveBroadcast',
+          'meeting',
+          'video',
+        ];
+
+        if (typeIncoming) {
+          if (onlyActivityTypes.includes(typeIncoming)) {
+            console.debug(typeIncoming, 'is in', onlyActivityTypes);
+            setActivityType(typeIncoming);
+            setActivityContentType(null);
+          } else {
+            console.debug(typeIncoming, 'is not in', onlyActivityTypes);
+            setActivityType(null);
+            setActivityContentType(typeIncoming as ActivitySubTypeValueType);
+          }
+        }
       } catch (e) {
         console.error(e);
       } finally {
